test commit

+++ root/prodamus/main.py
@@ -15,7 +15,7 @@ logging.basicConfig(
 
 async def confirm_payment(signature, data):
     # if verify_signature(signature, data):
-    if int(data['sum']) >= 2000 and data['payment_status'] == 'success':
+    if float(data['sum']) >= 2000 and data['payment_status'] == 'success':
         logging.info('we are in verify')
         session = db.Session()
         user = session.query(models.User).filter(models.User.order_id == data['order_num']).first()
